Add new TypeScript type definitions for controller and operation objects

+++ lockin.js
@@ -1,49 +1,48 @@
 const fs = require("fs");
 const path = require("path");
-
+const { compile } = require("json-schema-to-typescript");
 const schemas_directory_path = path.join(__dirname, "source/schemas");
-const main_file_path = path.join(__dirname, "source/index.js");
 
 async function lockin() {
-  const schemas = fs
-    .readdirSync(schemas_directory_path)
+  const schemas = fs.readdirSync(schemas_directory_path);
 
   const symbols = [];
 
-  fs.unlinkSync(main_file_path);
+  schemas
+    .filter((schema_file) => schema_file.endsWith(".json"))
+    .forEach(async (schema_file, index) => {
+      const name = schema_file.split("-").at(0);
 
-  schemas.forEach(async (schema_file, index) => {
-    const name = schema_file.split("-").at(0)
+      symbols.push([name, schema_file]);
 
-    const schema_path = path.join(schemas_directory_path, schema_file);
+      const main_file_path = path.join(__dirname, "types", `${name}.d.ts`);
+      if (fs.existsSync(main_file_path)) {
+        fs.unlinkSync(main_file_path);
+      }
 
-    const schema = require(schema_path);
+      const schema_path = path.join(schemas_directory_path, schema_file);
 
-    const type = `
+      const schema = require(schema_path);
 
-    const ${name} = /** @type {const} @satisfies {import('json-schema-to-ts').JSONSchema} */(${JSON.stringify(
-      schema,
-      null,
-      2
-    )})
+      schema.$id = name;
 
-    /**
-     * ${name}
-     * @typedef {import('json-schema-to-ts').FromSchema<typeof ${name}>} ${name}Type
-     */
+      try {
+        const type = await compile(schema, name);
+        fs.writeFileSync(main_file_path, type);
+      } catch (error) {
+        console.log("name: ", name, "err: ", error);
+      }
+    });
 
-    `;
+  const index_file_path = path.join(__dirname, "source", "index.js");
 
-    fs.appendFileSync(main_file_path, type);
+  let index = symbols
+    .map(([name, path]) => `const ${name} = require("./schemas/${path}")`)
+    .join("\n");
 
-    symbols.push(name);
-  });
+  index += `\nmodule.exports = {${symbols.map(([name]) => name).join(", ")}}`;
 
-  const exports = `
-    module.exports = { ${symbols.join(",")}}
-  `;
-
-  fs.appendFileSync(main_file_path, exports);
+  fs.writeFileSync(index_file_path, index);
 }
 
 lockin();
